Reorg Book handlers
