+++ Source/ACE.Server/WorldObjects/WorldObject.cs
@@ -464,24 +464,6 @@ namespace ACE.Server.WorldObjects
             }
         }
 
-        public void ReadBookPage(Session reader, uint pageNum)
-        {
-            //PageData pageData = new PageData();
-            //AceObjectPropertiesBook bookPage = PropertiesBook[pageNum];
-
-            //pageData.AuthorID = bookPage.AuthorId;
-            //pageData.AuthorName = bookPage.AuthorName;
-            //pageData.AuthorAccount = bookPage.AuthorAccount;
-            //pageData.PageIdx = pageNum;
-            //pageData.PageText = bookPage.PageText;
-            //pageData.IgnoreAuthor = false;
-            //// TODO - check for PropertyBool.IgnoreAuthor flag
-
-            //var bookDataResponse = new GameEventBookPageDataResponse(reader, Guid.Full, pageData);
-            //reader.Network.EnqueueSend(bookDataResponse);
-        }
-
- 
         public string DebugOutputString(WorldObject obj)
         {
             var sb = new StringBuilder();
